applications/workflows/requirements.txt

Flask
PyYAML
boto3
chevron
dropbox
fitbit
#git+https://gitlab.com/mcfitz2/workflows.git#egg=common&subdirectory=common-libs/python
httpx
ijson
imapclient
lxml
mariadb
peewee
phonenumbers
prefect
prefect-aws
psycopg2-binary
pymongo
python-dateutil
python-dotenv
pytz
requests
stravalib
tqdm
xmltodict
bs4
oauthlib
paramiko
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
